+++ actix-http/CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Minimum supported Rust version (MSRV) is now 1.52.
 
 
 ## 3.0.0-beta.10 - 2021-09-09