+++ actix-http/CHANGES.md
@@ -1,7 +1,11 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+### Changed
+* Updated rustls to v0.20. [#2414]
 * Minimum supported Rust version (MSRV) is now 1.52.
+
+[#2414]: https://github.com/actix/actix-web/pull/2414
 
 
 ## 3.0.0-beta.10 - 2021-09-09